requirements2.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
distlib==0.3.8
Django==5.1.1
django-admin-rangefilter==0.8.3
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-filters==0.11.1
filelock==3.16.1
packaging==24.1
pillow==11.1.0
pipenv==2024.0.3
platformdirs==4.3.6
psycopg2-binary==2.9.9
reportlab==4.2.5
setuptools==75.1.0
sqlparse==0.5.1
virtualenv==20.26.5
